+++ noodles-fastq/src/fai/record/field.rs
@@ -1,8 +1,8 @@
 /// A FASTA index record field.
 #[derive(Clone, Copy, Debug, Eq, PartialEq)]
 pub enum Field {
-    /// The read name.
-    ReadName,
+    /// The name.
+    Name,
     /// The total length of the sequence.
     Length,
     /// The offset to the sequence from the start.